vectorstore: Fix the vectors synchronization issue
After the vectors are created by Chroma.from_documents, the embedding
vectors are temporarily saved in the memory and would need to call the
persist() API to write back the cached value into the disk storage by
the __del__ function, which will be called at the end of the object life
cycle.

    def __del__(self):
        self.db.persist()

Actually, the object life is not predictable; for example, when people
finish the program by ctrl + C, it appears that the object would not be
written back by __del__(). As a result, it will be better to write back
the data immediately after the vectors construction. Or we can implement
this 'with' semantics to limit the object's life cycle. For example:

class MyClass:
    def __init__(self):
        self.db = open_database()

    def __enter__(self):
        return self

    def __exit__(self, exc_type, exc_value, traceback):
        if self.db is not None:
            self.db.persist()
            self.db.close()

    def do_something(self):
        # perform some operation with the database
        pass

with MyClass() as obj:
    obj.do_something()

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>